Improve error handling
Previously, an error in the word list was fatal, unhelpful, and leaky. I
added a message saying which line was responsible for the failure, and
then freed the word list struct array. Hopefully there are no dumb
errors, but it's totally possible. I'll be honest, I was a little
reckless with how I freed the memory here, although I'm mostly confident
in the implementation.